1. Avoid precision errors in stringification of transform list values.    Thanks Nick!
PR 40336


git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/batik/trunk@498484 13f79535-47bb-0310-9956-ffa450edef68

+++ sources/org/apache/batik/dom/svg/AbstractSVGTransformList.java
@@ -444,10 +444,10 @@ public abstract class AbstractSVGTransformList
             switch(type) {
                 case SVGTransform.SVG_TRANSFORM_TRANSLATE:
                     buf.append("translate(");
-                    buf.append(affineTransform.getTranslateX());
+                    buf.append((float) affineTransform.getTranslateX());
                     if (!xOnly) {
                         buf.append(' ');
-                        buf.append(affineTransform.getTranslateY());
+                        buf.append((float) affineTransform.getTranslateY());
                     }
                     buf.append(')');
                     break;
@@ -464,10 +464,10 @@ public abstract class AbstractSVGTransformList
                     break;
                 case SVGTransform.SVG_TRANSFORM_SCALE:
                     buf.append("scale(");
-                    buf.append(affineTransform.getScaleX());
+                    buf.append((float) affineTransform.getScaleX());
                     if (!xOnly) {
                         buf.append(' ');
-                        buf.append(affineTransform.getScaleY());
+                        buf.append((float) affineTransform.getScaleY());
                     }
                     buf.append(')');
                     break;
